Psychtoolbox for Linux: Add support for 64-Bit Octave on 64-Bit Linux.
This adds 64-Bit mex files for 64 bit Octave on Linux.
Build script updated, some compiler warnings fixed,
M-Files of installer/updater etc. modified to allow
64 bit Octave on Linux.

=> New subfolder PsychBasic/Octave3LinuxFiles64/ housese
   all 64 bit mex files for Octave 64 on Linux.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@2154 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/Base/PsychConstants.h
@@ -142,7 +142,7 @@ typedef unsigned char		psych_bool;
 
 //abstract up simple data types. 
 #if PSYCH_SYSTEM == PSYCH_LINUX
-        typedef long long						psych_int64;
+        typedef int64_t                         psych_int64;
         typedef unsigned long long              psych_uint64;
         typedef unsigned int                    psych_uint32;
         typedef unsigned char                   psych_uint8;
